Single family home neighborhoods are no longer middle class
Per
https://www.pewresearch.org/fact-tank/2018/09/06/the-american-middle-class-is-stable-in-size-but-losing-ground-financially-to-upper-income-families/,
the middle class is adults making 2/3 to double the US median
household income, a range from $45,200 to $135,600.

Per slide 5 of the California Association of Realtors presentation
at https://www.car.org/en/marketdata/data/haitraditional, the income
required to afford the median single family home in California is
$122,340. This is at the very high end of the middle income range.
It's also a stark contrast with the income required to afford the
median US single family home, $55,850.

Given these statistics and income requirements I don't think it's fair
to categorize single family home neighborhoods as "middle class." The
residents are either upper class (if they are moving in now) or have
built a significant amount of equity in their homes.